реализация запроса и отображение данных

+++ lib/models/weather_forecast_daily.dart
@@ -141,11 +141,11 @@ class WeatherList {
         weather!.add(new Weather.fromJson(v));
       });
     }
-    speed = json['speed'];
+    speed = json['speed'].toDouble();
     deg = json['deg'];
-    gust = json['gust'];
+    gust = json['gust'].toDouble();
     clouds = json['clouds'];
-    pop = json['pop'];
+    pop = json['pop'].toDouble();
     rain = json['rain'];
   }
 
@@ -215,10 +215,10 @@ class FeelsLike {
   FeelsLike({this.day, this.night, this.eve, this.morn});
 
   FeelsLike.fromJson(Map<String, dynamic> json) {
-    day = json['day'];
-    night = json['night'];
-    eve = json['eve'];
-    morn = json['morn'];
+    day = json['day'].toDouble();
+    night = json['night'].toDouble();
+    eve = json['eve'].toDouble();
+    morn = json['morn'].toDouble();
   }
 
   Map<String, dynamic> toJson() {
